+++ src/interfaces/index.ts
@@ -1,5 +1,7 @@
 export * from './network-change.message';
 export * from './request-handler.interface';
+export * from './request-options.interface';
+export * from './respond-options.interface';
 export * from './serializer.interface';
 export * from './subscribe-handler.interface';
 export * from './subscription.interface';